move select logic to WardrobeView
